+++ src/actions/genres/remove-genre.js
